cargo, lib, profile: remove log dependency
There is a single call to the info!() macro in the code base for a field
that is being set on the returned Profile struct.  It is more useful
for consumers to log the profile's path if they are interested in it.

This removes the log crate as a dependency.